requirements.txt
google-api-python-client>=2.148.0,<3
google-auth-httplib2>=0.2.0,<1
google-auth-oauthlib>=1.2.0,<2
